requirements-production.txt
# Production requirements for DSP Eco Tracker
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Caching==2.1.0
Flask-Limiter==3.5.0

# Database
PyMySQL==1.1.0
SQLAlchemy==2.0.21
redis==5.0.1

# Machine Learning
scikit-learn==1.3.1
xgboost==1.7.6
pandas==2.1.1
numpy==1.24.3
joblib==1.3.2

# Web Scraping
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0

# Production Server
gunicorn==21.2.0
gevent==23.9.1

# Utilities
python-dotenv==1.0.0
pgeocode==0.4.0

# Monitoring & Logging
prometheus-client==0.17.1
structlog==23.2.0

# Security
bcrypt==4.0.1
cryptography==41.0.7